app/requirements.txt
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
Django==5.1.3
djangorestframework==3.15.2
filelock==3.16.1
flower==2.0.1
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
kombu==5.4.2
packaging==24.2
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tornado==6.4.2
tzdata==2024.2
vine==5.1.0
watchfiles==1.0.0
wcwidth==0.2.13
